pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scid-agent-showcase"
version = "0.1.0"
description = "Public showcase repo for SCID orchestration and replay architecture."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.21",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "packages*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
